// src/modules/sdk-flex-card/elements/view/layout/layout-footer/types/layout-footer.ts
import { AnyObject } from "../../../../../../../shared/types/common/AnyObject";
import { ElementType } from "../../../../core/types/element";

export type LayoutFooterType = {
  elementType: "grid";
  element: "layout.footer";
  version: number;
  childrens?: ElementType[];
  styles?: AnyObject;
};